just need data
setup for measuring the force and velocity ofistonic contractions. The afterload is supported by ashen prior to activation of the muscle and thus it is notelet by the muscle unit allslack in theolin is taken up by muscle shorting. When the size of the afterloads increased
Human foot motion during stance can be described as a smooth progression between four phases:  heel strike,  flat foot,  heal lift, and  toe push-off. The model and proposed optimization methods are focused on the characteristics of the forefoot since the push-off phase and heel lift require the most energy during the gait cycle.
If we assume the leg below the knee joint as one segment (no active joints), the kinetics is defined by three second-order equations, however, with the flexible geometry that introduces the ankle rotation:
[rowOfMax, colOfMax] = find(pressureImage == maxPressure);
In order to justify the reasons for this research, we review shortly the state of the art that dominated recent literature. New designs and materials in prosthetic foot-ankle assemblies offer the individual with lower limb amputation wide choice among the prostheses with rather similar characteristic, which complicates the task of the prosthetic clinic team in formulating the most appropriate prescription foThe most important clinical distinction is related to the internal design characteristics, which enable the component to simulate some actions of the human foot where findings relate to the inital vecociy of shorting mussel 
r each patient
The segment kinematics need to be estimated. In our case, this was done by a stereometric system of cameras. The ground reaction forces have been estimated using the force platform. The parameters required for the optimization are also the horizontal and vertical distances from the center of mass (CM) and the position of the ground reaction force (COP)